reorganize pull commands
